refactor: hierarchy separation
