+++ index.html
@@ -97,7 +97,11 @@
                 name="section-station"
                 id="section-station-selector"
               ></select>
-              <input type="number" id="section-order-input" />
+              <input
+                type="number"
+                id="section-order-input"
+                placeholder="순서"
+              />
               <button id="section-add-button">등록</button>
             </div>
             <div id="section-table-container">
